moved ros-model specific functionality in a new module
this is a temp solution. Might be moved to ros-model-parser
